requirements.txt
rpi_ws281x
adafruit-circuitpython-neopixel
RPi.GPIO
pika
